Remove (and replace when needed) @psalm-taint-source annotations
The way they are used in the EasyDB codebase seems to be incorrect, it
looks they should be "taint sinks" instead.

For example I would expect the following example to flagged as insecure:
```php
<?php
$db = \ParagonIE\EasyDB\Factory::fromArray([
   'mysql:host=localhost;dbname=something',
   'username',
   'putastrongpasswordhere'
]);
$statement = \ParagonIE\EasyDB\EasyStatement::open()
    ->with('last_login IS NOT NULL')
    ->orWith('email = ' . $_POST['search']);
$user = $db->single("SELECT * FROM users WHERE $statement", $statement->values());
```

but with the current annotations it is considered fine.

`@psalm-taint-source` does not work on a specific parameter of a method
but on the returned value [0]. This PR only make sure the sinks are
properly defined but it could be interesting to mark the methods
returning something the DB in the EasyDB class as an input source.

[0] https://psalm.dev/docs/security_analysis/custom_taint_sources/